+++ unstructured/ingest/main.py
@@ -162,12 +162,6 @@ class MainProcess:
     help="Path to the location in the local file system that will be processed.",
 )
 @click.option(
-    "--local-recursive",
-    is_flag=True,
-    default=False,
-    help="Support recursive local file processing.",
-)
-@click.option(
     "--local-file-glob",
     default=None,
     help="A comma-separated list of file globs to limit which types of local files are accepted,"
@@ -177,7 +171,14 @@ class MainProcess:
     "--remote-url",
     default=None,
     help="Remote fsspec URL formatted as `protocol://dir/path`, it can contain both "
-    "a directory or a single file. Supported protocols are: `s3`, `s3a`, `abfs`, and `az`.",
+    "a directory or a single file. Supported protocols are: `gcs`, `gs`, `s3`, `s3a`, `abfs` "
+    "and `az`.",
+)
+@click.option(
+    "--gcs-token",
+    default=None,
+    help="Token used to access Google Cloud. GCSFS will attempt to use your default gcloud creds"
+    "or get creds from the google metadata service or fall back to anonymous access.",
 )
 @click.option(
     "--s3-anonymous",
@@ -210,13 +211,6 @@ class MainProcess:
     "--drive-service-account-key",
     default=None,
     help="Path to the Google Drive service account json file.",
-)
-@click.option(
-    "--drive-recursive",
-    is_flag=True,
-    default=False,
-    help="Recursively download files in folders from the Google Drive ID, "
-    "otherwise stop at the files in provided folder level.",
 )
 @click.option(
     "--drive-extension",
@@ -412,17 +406,26 @@ class MainProcess:
     show_default=True,
     help="Number of parallel processes to process docs in.",
 )
+@click.option(
+    "--recursive",
+    is_flag=True,
+    default=False,
+    help="Recursively download files in their respective folders"
+    "otherwise stop at the files in provided folder level."
+    " Supported protocols are: `gcs`, `gs`, `s3`, `s3a`, `abfs` "
+    "`az`, `google drive` and `local`.",
+)
 @click.option("-v", "--verbose", is_flag=True, default=False)
 def main(
     ctx,
     remote_url,
     s3_anonymous,
+    gcs_token,
     azure_account_name,
     azure_account_key,
     azure_connection_string,
     drive_id,
     drive_service_account_key,
-    drive_recursive,
     drive_extension,
     biomed_path,
     biomed_api_id,
@@ -457,6 +460,7 @@ def main(
     structured_output_dir,
     reprocess,
     num_processes,
+    recursive,
     verbose,
     metadata_include,
     metadata_exclude,
@@ -468,7 +472,6 @@ def main(
     partition_strategy,
     api_key,
     local_input_path,
-    local_recursive,
     local_file_glob,
     download_only,
 ):
@@ -580,7 +583,19 @@ def main(
                 standard_config=standard_config,
                 config=SimpleS3Config(
                     path=remote_url,
+                    recursive=recursive,
                     access_kwargs={"anon": s3_anonymous},
+                ),
+            )
+        elif protocol in ("gs", "gcs"):
+            from unstructured.ingest.connector.gcs import GcsConnector, SimpleGcsConfig
+
+            doc_connector = GcsConnector(  # type: ignore
+                standard_config=standard_config,
+                config=SimpleGcsConfig(
+                    path=remote_url,
+                    recursive=recursive,
+                    access_kwargs={"token": gcs_token},
                 ),
             )
         elif protocol in ("abfs", "az"):
@@ -602,14 +617,15 @@ def main(
                 standard_config=standard_config,
                 config=SimpleAzureBlobStorageConfig(
                     path=remote_url,
+                    recursive=recursive,
                     access_kwargs=access_kwargs,
                 ),
             )
         else:
             warnings.warn(
                 f"`fsspec` protocol {protocol} is not directly supported by `unstructured`,"
-                " so use it at your own risk. Supported protocols are `s3`, `s3a`, `abfs`,"
-                " and `az`.",
+                " so use it at your own risk. Supported protocols are `gcs`, `gs`, `s3`, `s3a`,"
+                "`abfs` and `az`.",
                 UserWarning,
             )
 
@@ -622,6 +638,7 @@ def main(
                 standard_config=standard_config,
                 config=SimpleFsspecConfig(
                     path=remote_url,
+                    recursive=recursive,
                 ),
             )
     elif github_url:
@@ -721,7 +738,7 @@ def main(
             config=SimpleGoogleDriveConfig(
                 drive_id=drive_id,
                 service_account_key=drive_service_account_key,
-                recursive=drive_recursive,
+                recursive=recursive,
                 extension=drive_extension,
             ),
         )
@@ -753,7 +770,7 @@ def main(
             standard_config=standard_config,
             config=SimpleLocalConfig(
                 input_path=local_input_path,
-                recursive=local_recursive,
+                recursive=recursive,
                 file_glob=local_file_glob,
             ),
         )